requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1